Initial Vision Test
This adds the ocr template for google vision and some book covers for testing purposes.

It also adds the output that was given from the ocr.

No tests yet, but that may be unnecessary given we may not end up using this. This is only to serve for a template for future reference if we determine we need it.