Added Object Localization
This change comes with a lot of reformatting. Lots of comments and white space was cleaned up and made more readable.

I also added google vision object localization detection and a function to call it from the server for testing. I also saved its output in a text file.

I also changed up routing and implemented blueprints for the vision tests.