Adds tests and Uses Matrices
Old version used to take in a direct image (usually a jpeg) and would scan that before outputting. Changed to instead use a image matrix using pytorch and scan that instead. Makes it easier to manipulate and feels faster.

Added exceptions testing for yolo image so we can guarantee the input is compatible with YOLO.

Added tests for the new yolo function.

I also added a context manager to the yolo function to make it more consistent when testing the whole back end vs just this file vs using pytest on it.