Koyeb Server ready
Redo the commit to get Koyeb to work.
Removed YOLO_Test and it's respective files from the project as it was causing issues with deploying.
Removed some images in the vision folder that are not neccessary
Set up hierarchy for gunicorn
Added a requirements.txt
Updated README and added images